added the math library, therefore added the math compiler flag

+++ inc/con_lib.h
@@ -4,6 +4,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <stdbool.h>
+#include <math.h>
 
 typedef enum e_status
 {
@@ -246,7 +247,7 @@ t_obj	*ft_get_nearest_resource(t_obj *unit);
  */
 t_unit_config	*ft_get_unit_config(t_unit_type type);
 // ---------------- get utils ----------------
-long	ft_distance(t_obj *obj1, t_obj *obj2);
+double	ft_distance(t_obj *obj1, t_obj *obj2);
 
 // --------------- actions.c ---------------
 /**
